Convert entity state reducers in example app to functions.
AOT compiling does not work properly when using createReducer for
an AE entity's reducer.  The examples in the From Scratch section
recognize this and export a reducer function.  However the code in
the example app uses `createReducer` which causes a build failure
if using AOT compiling.  This commit updates the example application
documents to match the From Scratch documents when creating reducers.